chore(core): pr fixes
