+++ packages/core/src/config-manager/admin/routes/RegisterConfigRoute.ts
@@ -1,12 +1,7 @@
-import ConduitGrpcSdk, {
-  ConduitError,
-  ConduitRouteActions,
-  ConduitRouteParameters,
-} from '@conduitplatform/grpc-sdk';
-import { ConduitRoute, ConduitRouteReturnDefinition } from '@conduitplatform/hermes';
+import ConduitGrpcSdk, { ConduitRouteActions } from '@conduitplatform/grpc-sdk';
 import { ConduitCommons } from '@conduitplatform/commons';
-import getConfigRouteHandler from './GetConfigRouteHandler';
-import setConfigRouteHandler from './SetConfigRouteHandler';
+import getConfigRoute from './GetConfigRoute';
+import setConfigRoute from './SetConfigRoute';
 
 export function registerConfigRoute(
   grpcSdk: ConduitGrpcSdk,
@@ -15,35 +10,9 @@ export function registerConfigRoute(
   configSchema: any,
   routeAction: ConduitRouteActions.GET | ConduitRouteActions.PATCH,
 ) {
-  return new ConduitRoute(
-    {
-      path: `/config/${moduleName}`,
-      action: routeAction,
-      ...(routeAction === ConduitRouteActions.PATCH && {
-        bodyParams: {
-          config: { type: configSchema, required: true },
-        },
-      }),
-    },
-    new ConduitRouteReturnDefinition(
-      routeAction === ConduitRouteActions.GET
-        ? `Get${moduleName}ConfigRoute`
-        : `Set${moduleName}ConfigRoute`,
-      {
-        config: configSchema,
-      },
-    ),
-    async (params: ConduitRouteParameters) => {
-      let response;
-      routeAction === ConduitRouteActions.GET
-        ? (response = await getConfigRouteHandler(grpcSdk, moduleName))
-        : (response = await setConfigRouteHandler(
-            params.params!.config,
-            moduleName,
-            grpcSdk,
-            conduit,
-          ));
-      return response;
-    },
-  );
+  let response;
+  routeAction === ConduitRouteActions.GET
+    ? (response = getConfigRoute(grpcSdk, moduleName, configSchema))
+    : (response = setConfigRoute(moduleName, grpcSdk, conduit, configSchema));
+  return response;
 }
